[DITP-pilotage/pilote-2] feature: sépare Responsables et Accès en deux onglets, navigation par search params
La fiche collection passe à quatre onglets : Détails, Indicateurs, Responsables,
Accès. `CollectionUtilisateurs` est scindé en deux composants ; le bandeau de
verrouillage PROD, qui aurait été en triple exemplaire, devient
`ProdEditSectionHeader`. Le hook `useProdEditUnlock` reste appelé par chaque
parent, qui a besoin de `locked` pour désactiver ses contrôles.

L'onglet actif passe des `useState` locaux aux search params TanStack
(`?onglet=…`), comme sur la webapp : les trois écrans à onglets de l'admin —
collection, clé API, utilisateur — sont alignés, et un onglet devient
partageable et résistant au rechargement.

Espacement des sections repris en `flex flex-col gap-4` plutôt qu'en marges :
les marges de frères adjacents fusionnaient, et les classes utilitaires
introduites uniquement par ces fichiers neufs n'étaient pas générées par
Tailwind sans redémarrage du serveur de dev.

Refs:
Project: pilote-2
Client: DITP-pilotage
Category: feature
Stack: typescript, react, tailwind
Tags: admin, navigation
Files-changed: 8
