+++ apps/kpilote-admin/src/routes/_authed/utilisateurs/$id.tsx
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient, useSuspenseQuery } from '@tanstack/react-query'
 import { createFileRoute, Link, useNavigate } from '@tanstack/react-router'
-import { useState } from 'react'
+import { z } from 'zod'
 
 import { updateUtilisateur } from '@/api/utilisateurs'
 import { Breadcrumb } from '@/components/Breadcrumb'
@@ -14,7 +14,12 @@ import { indicateursAllQueryOptions } from '@/queries/indicateurs'
 import { principalPermissionsQueryOptions } from '@/queries/permissions'
 import { utilisateurQueryOptions } from '@/queries/utilisateurs'
 
+const searchSchema = z.object({
+  onglet: z.enum(['identite', 'permissions']).default('identite'),
+})
+
 export const Route = createFileRoute('/_authed/utilisateurs/$id')({
+  validateSearch: searchSchema,
   loader: async ({ context, params }) => {
     await Promise.all([
       context.queryClient.ensureQueryData(utilisateurQueryOptions(params.id)),
@@ -29,7 +34,7 @@ function EditUtilisateurComponent() {
   const { id } = Route.useParams()
   const navigate = useNavigate()
   const queryClient = useQueryClient()
-  const [tab, setTab] = useState<'identite' | 'permissions'>('identite')
+  const { onglet } = Route.useSearch()
 
   const { data: utilisateur } = useSuspenseQuery(utilisateurQueryOptions(id))
 
@@ -65,14 +70,24 @@ function EditUtilisateurComponent() {
       </Breadcrumb>
       <PageHeading title="Modifier l'utilisateur" />
 
-      <Tabs value={tab} onValueChange={(key) => setTab(key as 'identite' | 'permissions')}>
+      <Tabs
+        value={onglet}
+        onValueChange={(valeur) => {
+          void navigate({
+            to: '/utilisateurs/$id',
+            params: { id },
+            search: { onglet: valeur as typeof onglet },
+            replace: true,
+          })
+        }}
+      >
         <TabsList className="mb-6">
           <TabsTrigger value="identite">Identité</TabsTrigger>
           <TabsTrigger value="permissions">Permissions</TabsTrigger>
         </TabsList>
       </Tabs>
 
-      {tab === 'identite' ? (
+      {onglet === 'identite' ? (
         utilisateur ? (
           <UtilisateurForm
             mode="update"
